+++ cloud/custom/centos/install-common.sh
@@ -11,10 +11,9 @@ fi
 yum install -y -q zip unzip bzip2 xz tig jq sshpass screen python2-httpie wget vim moreutils
 
 yum install -y -q yum-utils
-yum-config-manager --add-repo http://download.opensuse.org/repositories/shells:/fish:/release:/3/RHEL_7/shells:fish:release:3.repo
-yum install -y -q fish
-curl -sL https://copr.fedorainfracloud.org/coprs/outman/emacs/repo/epel-7/outman-emacs-epel-7.repo -o /etc/yum.repos.d/emacs-copr.repo
-curl -sL https://copr.fedorainfracloud.org/coprs/carlwgeorge/ripgrep/repo/epel-7/carlwgeorge-ripgrep-epel-7.repo -o /etc/yum.repos.d/ripgrep.repo
+if ! yum repolist all | grep -s -q "^shells_fish_release_3"; then
+    yum install -y -q fish
+fi
 
 if [[ -n ${CUSTOM_UPDATE} && "${CUSTOM_UPDATE}" == "true" ]]; then
     yum -y -q update
